added a bunch of functions
working to put functions in files once they are done.  Also have
everything wrapped up in a players object that is a list of player, each
of which has a player specific graph. Also added/switched to
cities_short and edges_short, cutting all_asp cal time from 15 sec to
0.5 sec